+++ actionpack/test/controller/parameters/accessors_test.rb
@@ -203,6 +203,16 @@ class ParametersAccessorsTest < ActiveSupport::TestCase
     assert_not_predicate @params.transform_keys { |k| k }, :permitted?
   end
 
+  test "transform_keys without a block returns an enumerator" do
+    assert_kind_of Enumerator, @params.transform_keys
+    assert_kind_of ActionController::Parameters, @params.transform_keys.each { |k| k }
+  end
+
+  test "transform_keys! without a block returns an enumerator" do
+    assert_kind_of Enumerator, @params.transform_keys!
+    assert_kind_of ActionController::Parameters, @params.transform_keys!.each { |k| k }
+  end
+
   test "transform_values retains permitted status" do
     @params.permit!
     assert_predicate @params.transform_values { |v| v }, :permitted?
@@ -219,8 +229,9 @@ class ParametersAccessorsTest < ActiveSupport::TestCase
     end
   end
 
-  test "transform_values without block yieds an enumerator" do
+  test "transform_values without a block returns an enumerator" do
     assert_kind_of Enumerator, @params.transform_values
+    assert_kind_of ActionController::Parameters, @params.transform_values.each { |v| v }
   end
 
   test "transform_values! converts hashes to parameters" do
@@ -229,8 +240,9 @@ class ParametersAccessorsTest < ActiveSupport::TestCase
     end
   end
 
-  test "transform_values! without block yields an enumerator" do
+  test "transform_values! without a block returns an enumerator" do
     assert_kind_of Enumerator, @params.transform_values!
+    assert_kind_of ActionController::Parameters, @params.transform_values!.each { |v| v }
   end
 
   test "value? returns true if the given value is present in the params" do